Repo Refresh

+++ src/main/java/app/virtual_games/sudoku/handlers/HintButtonHandler.java
@@ -7,36 +7,36 @@ import javafx.event.EventHandler;
 
 
 /**
- * 
+ *
  * Custom event handler for the hint button click event.
- * 
+ *
  * @author Corey Caskey
  * @version 0.0.1
  *
  */
-public class HintButtonHandler implements EventHandler<ActionEvent> 
+public class HintButtonHandler implements EventHandler<ActionEvent>
 {
-	/**
-	 * 
-	 * Loads the hint into the sudoku puzzle and displays the time penalty.
-	 * 
-	 * @param clickEvent : hint button click event
-	 * 
-	 */
-	@Override
-	public void handle(ActionEvent clickEvent) 
-	{
-		int timePenalty = GameController.getTimePenalty();
-		
-		GameController.unhighlightSudokuPuzzle();
-		GameController.unclickNumberButton();
-		GameController.setCurrentClickedNumberButton(null);
-		GameController.hideTimePenalty();
-		GameController.stopTimePenaltyTimer();
-		GameController.showTimePenalty(timePenalty);
-		GameController.startTimePenaltyTimer();
-		GameController.incrementPlayingTime(timePenalty * 1000);
-		GameController.updateTimerLabel();
-		GameController.getHint();
-	}
+  /**
+   *
+   * Loads the hint into the sudoku puzzle and displays the time penalty.
+   *
+   * @param clickEvent : hint button click event
+   *
+   */
+  @Override
+  public void handle(ActionEvent clickEvent)
+  {
+    int timePenalty = GameController.getTimePenalty();
+
+    GameController.unhighlightSudokuPuzzle();
+    GameController.unclickNumberButton();
+    GameController.setCurrentClickedNumberButton(null);
+    GameController.hideTimePenalty();
+    GameController.stopTimePenaltyTimer();
+    GameController.showTimePenalty(timePenalty);
+    GameController.startTimePenaltyTimer();
+    GameController.incrementPlayingTime(timePenalty * ((long) 1000));
+    GameController.updateTimerLabel();
+    GameController.getHint();
+  }
 }
